Add script to act as a scratchpad, add to .zshrc
This commit adds a small shell script to act as a vim scratchpad,
creating a temporary file in a temporary directory. It takes one
parameter, being the extension. This will enable syntax highlighting in
Vim.

The scripts folder is intended to be placed in the HOME directory,
hidden in a directory called .programs.

ln -s ~/.dotfiles/scripts ~/.programs

If this directory exists, it will be added to $PATH.

Also, if .shortcuts and .localrc do not exist, they won't be sourced.